refactor: merge layout tag checks into tags.engine
Move check_is_xor/reloc/code from the deleted checks module into engine
and update layout/base imports. Keeps layout-derived tagging alongside
database tagger wiring in one module.
