+++ floss/layout/base.py
@@ -25,7 +25,7 @@ import pefile
 from pydantic import Field, BaseModel, ConfigDict
 
 from floss.ranges import Range, Slice, OffsetRanges
-from floss.tags.checks import check_is_xor, check_is_code, check_is_reloc
+from floss.tags.engine import check_is_xor, check_is_code, check_is_reloc
 from floss.layout.types import Tag, TaggedString, ExtractedString
 
 Tagger = Callable[[ExtractedString], Sequence[Tag]]
